Stuff
+NoRender
+NoRenderFog
#FixedNoJitter

+++ src/main/java/net/minecraft/src/RenderList.java
@@ -26,9 +26,9 @@ public class RenderList {
         field_1242_a = i;
         field_1241_b = j;
         field_1240_c = k;
-        field_1239_d = (float) d;
-        field_1238_e = (float) d1;
-        field_1237_f = (float) d2;
+        field_1239_d = d;
+        field_1238_e = d1;
+        field_1237_f = d2;
     }
 
     public boolean func_862_a(int i, int j, int k) {
@@ -70,9 +70,9 @@ public class RenderList {
     private int field_1242_a;
     private int field_1241_b;
     private int field_1240_c;
-    private float field_1239_d;
-    private float field_1238_e;
-    private float field_1237_f;
+    private double field_1239_d;
+    private double field_1238_e;
+    private double field_1237_f;
     private IntBuffer field_1236_g;
     private boolean field_1235_h;
     private boolean field_1234_i;
